requirements.txt
flask==3.0.2
markdown==3.5.2
gunicorn==21.2.0
google-cloud-monitoring==2.19.0
google-cloud-logging==3.9.0
openai==1.12.0 